requirements.txt
numpy==1.26.4 
pandas==2.2.2 
scikit-learn==1.5.2
opencv-python==4.10.0.84 
mediapipe==0.10.14 
flask==3.0.3 
flask-cors==5.0.0 
werkzeug==3.0.3 
google-generativeai==0.8.3 
matplotlib==3.9.2 
seaborn==0.13.2